refactor: Merge types/docs from testing/auth/oauth_grant_handler.pyi
- Merged type hints and docstrings from crudclient/testing/auth/oauth_grant_handler.pyi into crudclient/testing/auth/oauth_grant_handler.py.

- Deleted crudclient/testing/auth/oauth_grant_handler.pyi.
