--- crudclient/testing/auth/oauth_grant_handler.pyi
@@ -1,152 +1,0 @@
-"""
-OAuth grant type handling utilities for testing.
-
-This module provides a class for handling different OAuth grant types,
-including authorization_code, client_credentials, password, and refresh_token.
-"""
-
-from datetime import datetime
-from typing import Any, Dict, Optional, Set
-
-from .oauth_scope_validator import OAuthScopeValidator
-from .oauth_token_manager import OAuthTokenManager
-
-class OAuthGrantHandler:
-    """
-    OAuth grant handler for processing different grant types.
-
-    This class provides methods for handling different OAuth grant types,
-    including authorization_code, client_credentials, password, and refresh_token.
-    """
-
-    token_manager: OAuthTokenManager
-    scope_validator: OAuthScopeValidator
-    grant_type: str
-    supported_grant_types: Set[str]
-
-    def __init__(self, token_manager: OAuthTokenManager, scope_validator: OAuthScopeValidator) -> None:
-        """
-        Initialize an OAuth grant handler.
-
-        Args:
-            token_manager: The token manager
-            scope_validator: The scope validator
-        """
-        ...
-
-    def set_supported_grant_types(self, grant_types: Set[str]) -> None:
-        """
-        Set the supported grant types.
-
-        Args:
-            grant_types: The supported grant types
-        """
-        ...
-
-    def set_default_grant_type(self, grant_type: str) -> None:
-        """
-        Set the default grant type.
-
-        Args:
-            grant_type: The default grant type
-
-        Raises:
-            ValueError: If the grant type is not supported
-        """
-        ...
-
-    def handle_token_request(
-        self,
-        grant_type: str,
-        client_id: str,
-        client_secret: str,
-        scope: Optional[str] = None,
-        code: Optional[str] = None,
-        redirect_uri: Optional[str] = None,
-        username: Optional[str] = None,
-        password: Optional[str] = None,
-        refresh_token: Optional[str] = None,
-    ) -> Dict[str, Any]:
-        """
-        Handle a token request based on the grant type.
-
-        Args:
-            grant_type: The OAuth grant type
-            client_id: The client ID
-            client_secret: The client secret
-            scope: The requested scope
-            code: The authorization code (for authorization_code grant)
-            redirect_uri: The redirect URI (for authorization_code grant)
-            username: The username (for password grant)
-            password: The password (for password grant)
-            refresh_token: The refresh token (for refresh_token grant)
-
-        Returns:
-            A dictionary with the token response or error
-        """
-        ...
-
-    def _handle_authorization_code_grant(
-        self, client_id: str, client_secret: str, code: Optional[str], redirect_uri: Optional[str], scope: Optional[str]
-    ) -> Dict[str, Any]:
-        """
-        Handle authorization_code grant type.
-
-        Args:
-            client_id: The client ID
-            client_secret: The client secret
-            code: The authorization code
-            redirect_uri: The redirect URI
-            scope: The requested scope
-
-        Returns:
-            A dictionary with the token response or error
-        """
-        ...
-
-    def _handle_client_credentials_grant(self, client_id: str, client_secret: str, scope: Optional[str]) -> Dict[str, Any]:
-        """
-        Handle client_credentials grant type.
-
-        Args:
-            client_id: The client ID
-            client_secret: The client secret
-            scope: The requested scope
-
-        Returns:
-            A dictionary with the token response or error
-        """
-        ...
-
-    def _handle_password_grant(
-        self, client_id: str, client_secret: str, username: Optional[str], password: Optional[str], scope: Optional[str]
-    ) -> Dict[str, Any]:
-        """
-        Handle password grant type.
-
-        Args:
-            client_id: The client ID
-            client_secret: The client secret
-            username: The username
-            password: The password
-            scope: The requested scope
-
-        Returns:
-            A dictionary with the token response or error
-        """
-        ...
-
-    def _handle_refresh_token_grant(self, client_id: str, client_secret: str, refresh_token: Optional[str], scope: Optional[str]) -> Dict[str, Any]:
-        """
-        Handle refresh_token grant type.
-
-        Args:
-            client_id: The client ID
-            client_secret: The client secret
-            refresh_token: The refresh token
-            scope: The requested scope
-
-        Returns:
-            A dictionary with the token response or error
-        """
-        ...
